+++ src/bootx64/src/uefi.rs
@@ -1,7 +1,7 @@
 // References
 // https://uefi.org/sites/default/files/resources/UEFI_Spec_2_9_2021_03_18.pdf
 
-pub mod protocol;
+pub mod protocols;
 pub mod system_table;
 pub mod table_header;
 pub mod types;
