+++ src/bootx64/src/uefi/protocols.rs
