+++ src/bootx64/src/uefi/protocols/simple_text_input.rs
